Updated Start Bootstrap Site Architecture
We've updated the Start Bootstrap site to PHP instead of using static
HTML files for every page. The templates are still HTML pages, but the
master site is now using PHP includes for the navigation, footer, some
meta tags, and JavaScript to allow for easier updating of the master
site.